refactor: AlertStyle 열거형 정의, AlertViewDataMaker 메서드 하나로 통합
